fix(offline-toast): move connectivity polling to App.svelte and fix fly/transform conflict
Polling was confined to Home.svelte so the isOnline store stopped updating
when the user navigated to other views. Moves pingConnectivity + the
visibilitychange-aware interval to App.svelte so it runs for the lifetime
of the app.

Also fixes the transition:fly + transform:translateX(-50%) centering
conflict on .offline-toast: Svelte's fly applies an inline transform that
overwrites the CSS centering transform, causing a horizontal jump during
the slide animation. Switches to margin-inline:auto / width:fit-content
centering so fly can animate without interference.
